chore(e2e): refactor services
