+++ core-web/libs/edit-content/src/lib/utils/mocks.ts
@@ -18,10 +18,13 @@ import {
 } from '@dotcms/dotcms-models';
 import { MockDotMessageService } from '@dotcms/utils-testing';
 
-import { CustomTreeNode, TreeNodeItem } from '../fields/dot-edit-content-host-folder-field/models/tree-item.model';
 import { WYSIWYG_MOCK } from '../fields/dot-edit-content-wysiwyg-field/mocks/dot-edit-content-wysiwyg-field.mock';
 import { FIELD_TYPES } from '../models/dot-edit-content-field.enum';
 import { DotFormData } from '../models/dot-edit-content-form.interface';
+import {
+    CustomTreeNode,
+    TreeNodeItem
+} from '../models/dot-edit-content-host-folder-field.interface';
 import { DotWorkflowState } from '../models/dot-edit-content.model';
 
 /* FIELDS MOCK BY TYPE */
